Separate the line chart into a separate view.
